refactor(test): extract the local-Supabase test harness into a shared helper (PUL-236)
The two encryption suites carried ~140 identical lines of stack
resolution (CLI path stripping, status parsing, health probe, env
fallback) — and the copies had already diverged once, which is exactly
what produced today's merge conflicts against preview. Verbatim move to
src/test/local-supabase.ts next to the existing test-mocks helper;
behavior unchanged, -179 net lines.
